backend/requirements.txt
fastapi==0.88.0
uvicorn[standard]==0.20.0
sqlalchemy==1.4.53
psycopg2-binary==2.9.5
alembic==1.8.1
python-dotenv==0.21.1
pydantic==1.10.2
python-multipart==0.0.5
gunicorn==20.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.23.3
